Revision of Angel's First App (Part 1)
### Key notes

1. It’s really important to tabulate your code correctly as it makes it
much easier to read and much quicker to identify errors
2. Given we’re using Devise for “Authentication” (to create user
accounts with passwords, etc.) we will not have a `new` or `create`
action in our UsersController
3. We should apply `before_action :authenticate_user!` to any actions
in our controllers that should only be accessible to signed in users.
This is a Devise method that redirects people to the log in page if
they try to access a page that is restricted to logged in users
4. We should only declare each resource once in our `routes.rb`
5. The `load_and_authorize_resource` method comes from the `CanCanCan`
gem, which we use for “Authorisation”. It applies the permissions
outlined in the `ability.rb` file that controls which users can take
which actions.
6. Just to re-affirm, we use Devise for “Authentication” (ensuring the
user is logged in with a password) and we use CanCanCan for
“Authorisation” (controlling which users can access and manage which
resources).

Great work Angel!